instance buyy (#38)

+++ ResiBuy.Server/Application/Commands/OrderCommands/CreateOrderCommandHandler.cs
@@ -25,9 +25,9 @@ namespace ResiBuy.Server.Application.Commands.OrderCommands
             if (user == null) throw new CustomException(ExceptionErrorCode.ValidationFailed, "Người dùng không tồn tại");
             if(!user.Roles.Contains(Constants.CustomerRole)) throw new CustomException(ExceptionErrorCode.ValidationFailed, "Chỉ có người mua mới tạo được đơn hàng.");
             if (user.Cart == null)
-                throw new CustomException(ExceptionErrorCode.ValidationFailed, "Giỏ hàng không có sản phẩm nào.");
-            var cart = await cartDbService.GetByIdAsync(user.Cart.Id);
-            if(!cart.CartItems.Any()) throw new CustomException(ExceptionErrorCode.ValidationFailed, "Giỏ hàng không có sản phẩm nào.");
+                throw new CustomException(ExceptionErrorCode.ValidationFailed, "Không có tồn tại giỏ hàng.");
+            var cart = await cartDbService.GetByIdAsync(user.Cart.Id); 
+            if(!cart.CartItems.Any() && !dto.IsInstance) throw new CustomException(ExceptionErrorCode.ValidationFailed, "Giỏ hàng không có sản phẩm nào.");
             if (!cart.IsCheckingOut)
                 throw new CustomException(ExceptionErrorCode.ValidationFailed, "Giỏ hàng chưa ở trạng thái thanh toán.");
             cart.IsCheckingOut = false;
@@ -42,6 +42,7 @@ namespace ResiBuy.Server.Application.Commands.OrderCommands
             var orders = dto.Orders.Select(o => new Order(o.Id, o.TotalPrice, dto.PaymentMethod, o.Note, dto.AddressId, dto.UserId, o.StoreId, o.Items.Select(i => new OrderItem(i.Quantity, i.Price, o.Id, i.ProductDetailId)).ToList()));
             var createdOrders = await orderDbService.CreateBatchAsync(orders);
             if(createdOrders == null || !createdOrders.Any()) throw new CustomException(ExceptionErrorCode.CreateFailed, "Không thể tạo đơn hàng");
+            if(!dto.IsInstance)
              await cartItemDbService.DeleteBatchByProductDetailIdAsync(cart.Id , createdOrders.SelectMany(o => o.Items).Select(ci => ci.ProductDetailId));
             var notiUserIds = createdOrders.Select(o => o.StoreId.ToString()).Append(user.Id).Distinct().ToList();
             notificationService.SendNotification("OrderCreated", createdOrders, "", notiUserIds);
